ui(perps): inline max-leverage input in the stats row; drop the panel
Per review: no boxed panel. The See info 'Maximum leverage' row itself
becomes the admin control - current value plus a bare number input and
Set button in the value cell (purple admin row, unresolved perps only).
perp-admin-panel.tsx is now unreferenced and deleted.
