+++ web/components/contract/contract-info-dialog.tsx
@@ -36,7 +36,6 @@ import { InfoBox } from '../widgets/info-box'
 import { InfoTooltip } from '../widgets/info-tooltip'
 import ShortToggle from '../widgets/short-toggle'
 import { linkClass } from '../widgets/site-link'
-import { PerpAdminPanel } from '../perps/perp-admin-panel'
 import { Table } from '../widgets/table'
 import { ContractHistoryButton } from './contract-edit-history-button'
 
@@ -583,6 +582,7 @@ export const Stats = (props: {
 function PerpStatsRows(props: { contract: PerpContract }) {
   const { contract } = props
   const isAdmin = useAdmin()
+  const canEditLeverage = isAdmin && !contract.isResolved
   const price =
     contract.resolution === 'MKT'
       ? Number(contract.resolvedOraclePrice ?? contract.oraclePrice)
@@ -630,24 +630,18 @@ function PerpStatsRows(props: { contract: PerpContract }) {
           />
         </td>
       </tr>
-      {isAdmin && !contract.isResolved ? (
-        // Admins get the live max-leverage editor in place of the read-only
-        // row, alongside the dialog's other purple admin controls.
-        <tr className="bg-purple-500/30">
-          <td colSpan={2} className="whitespace-normal">
-            <PerpAdminPanel contract={contract} />
-          </td>
-        </tr>
-      ) : (
-        <tr>
-          <td>Maximum leverage</td>
-          <td>
-            {Number.isFinite(contract.maxLeverage)
-              ? `${formatWithCommas(contract.maxLeverage)}×`
-              : '—'}
-          </td>
-        </tr>
-      )}
+      <tr className={clsx(canEditLeverage && 'bg-purple-500/30')}>
+        <td>Maximum leverage</td>
+        <td>
+          {canEditLeverage ? (
+            <MaxLeverageInput contract={contract} />
+          ) : Number.isFinite(contract.maxLeverage) ? (
+            `${formatWithCommas(contract.maxLeverage)}×`
+          ) : (
+            '—'
+          )}
+        </td>
+      </tr>
       <tr>
         <td>
           Current funding{' '}
@@ -659,6 +653,76 @@ function PerpStatsRows(props: { contract: PerpContract }) {
         </td>
       </tr>
     </>
+  )
+}
+
+// Inline admin editor for update-perp-config. The change applies to the next
+// trade immediately (the engine re-reads the contract per trade); lowering
+// the cap only constrains new opens — existing positions are grandfathered.
+function MaxLeverageInput(props: { contract: PerpContract }) {
+  const { contract } = props
+  const [input, setInput] = useState('')
+  const [saving, setSaving] = useState(false)
+  // Bridge the gap between a save and the next contract poll so the row
+  // never flashes the pre-save cap.
+  const [justSaved, setJustSaved] = useState<number | null>(null)
+  const current =
+    justSaved != null && justSaved !== contract.maxLeverage
+      ? justSaved
+      : contract.maxLeverage
+  const parsed = Number(input)
+  const valid =
+    input !== '' && Number.isFinite(parsed) && parsed > 1 && parsed <= 100
+
+  const submit = async () => {
+    if (!valid || saving || parsed === current) return
+    setSaving(true)
+    try {
+      const res = await api('update-perp-config', {
+        contractId: contract.id,
+        maxLeverage: parsed,
+      })
+      setJustSaved(res.maxLeverage)
+      setInput('')
+      toast.success(`Max leverage is now ${res.maxLeverage}×`)
+    } catch (err) {
+      toast.error(
+        err instanceof Error ? err.message : 'Failed to update max leverage'
+      )
+    } finally {
+      setSaving(false)
+    }
+  }
+
+  return (
+    <Row className="items-center gap-2">
+      <span className="tabular-nums">
+        {Number.isFinite(current) ? `${formatWithCommas(current)}×` : '—'}
+      </span>
+      <input
+        type="number"
+        min={1}
+        max={100}
+        step={0.5}
+        value={input}
+        disabled={saving}
+        placeholder="New cap"
+        onChange={(e) => setInput(e.target.value)}
+        onKeyDown={(e) => {
+          if (e.key === 'Enter') submit()
+        }}
+        className="bg-canvas-0 border-ink-300 h-7 w-24 rounded-md border px-2 text-sm"
+      />
+      <Button
+        size="2xs"
+        color="indigo-outline"
+        disabled={!valid || saving || parsed === current}
+        loading={saving}
+        onClick={submit}
+      >
+        Set
+      </Button>
+    </Row>
   )
 }
